combined the update and create modals into a single "editable" modal
